[alphes] th09, th095:  Aya Portraits
Considering that Aya's portraits in both games are virtually the same, I
just went ahead and replaced the portraits in TH095 with the same
replacements as TH09. This also finishes TH095 support. As such, I added
TH095, TH125, and TH143 to the table of supported games as I plan to
support those as well.