Resized to Proper Size
Resized to Original size that the Touhou Toolkit wants in order to
recreate the th09.dat archive.

I have properly resized it so Touhou toolkit will be able to package it,
just in case anyone wants to build there own PoFV with Alphas.  Using
something like Touhou Toolkit, so you don't need Thcrap to use the Alpha
artwork.  I just couldn't get th09 to run with Thcrap and I'm in no way
computer illiterate, so I decided to bypass Thcrap Patcher and do it the
hard way.  Obviously use Thcrap if you can, but this is an alternative
way to go about it.

By the way NO changes have been made to the actual images, I've only
resized the canvas so they will be recognized by Touhou Toolkit as the
proper image sizes.  So no worries about it being better or worst
quality.  The quality has not been changed and the images have not been
resized, only the canvas behind the images have been resized and I went
though it image by image only for th09.

If anyone would like me to resize any of the other games images (if they
need to be), please let me know and I'll get right on it :)