[alphes] th06: Add Remilia and Flandre
With that, Touhou 6 support is basically done. Ideally, I would like to
also change the portraits on the character select screen, but due to the
way the images are split and because translatable text are on those, it
would have to be language dependent which is a pain. Thus, I'll just
declare those out of scope and say it is finished.